test: move L1 tests to a separate module
